+++ src/main/java/loanbook/model/loan/Email.java
@@ -38,15 +38,7 @@ public class Email extends DataField<String> implements Censor {
      * Check if user's new {@code Email} is a valid Gmail.
      */
     public static boolean isValidGmail(String email) {
-        int index = email.indexOf('@');
-
-        if (index == -1) {
-            return false;
-        }
-
-        String domain = email.substring(index + 1);
-
-        return domain.length() == 9 && domain.equals("gmail.com");
+        return email.matches("^[a-z0-9](\\.?[a-zA-Z0-9])+@gmail\\.com$");
     }
 
     /**
