+++ src/test/java/loanbook/testutil/TypicalLoanBook.java
@@ -29,4 +29,23 @@ public class TypicalLoanBook {
 
         return lb;
     }
+
+    /**
+     * Returns an {@code LoanBook} with all the bikes and loans only for the remind tests.
+     */
+    public static LoanBook getLoanBookForRemindTest() {
+        LoanBook lb = new LoanBook();
+        for (Bike bike : TypicalBikes.getTypicalBikes()) {
+            lb.addBike(bike);
+        }
+        for (Loan loan : TypicalLoans.getLoansForRemindTest()) {
+            lb.addLoan(loan);
+        }
+
+        int lastUsedId = lb.getLoanList().size() + LoanId.MINIMUM_ID - 1;
+        LoanId lastUsedLoanId = LoanId.isValidLoanId(lastUsedId) ? LoanId.fromInt(lastUsedId) : null;
+        lb.setLoanIdManager(new LoanIdManager(lastUsedLoanId));
+
+        return lb;
+    }
 }
